+++ Ext/Ktx/src/main/java/dev/testify/internal/helpers/IsRunningOnUiThread.kt
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2022 ndtp
+ * Copyright (c) 2023-2025 ndtp
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,15 +21,16 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
+package dev.testify.internal.helpers
 
-package dev.testify
-
-import android.app.Activity
-import android.graphics.Bitmap
-import android.view.View
+import android.os.Looper
+import dev.testify.internal.annotation.ExcludeFromJacocoGeneratedReport
 
 /**
- * Returns a [Bitmap] from the provided [Activity] and [View].
- * Invoked from the UI thread.
+ * Helper extension method to determine if the current thread is the UI thread.
+ *
+ * @return True if the current thread is the UI thread, false otherwise.
  */
-typealias CaptureMethod = (activity: Activity, targetView: View?) -> Bitmap?
+@ExcludeFromJacocoGeneratedReport
+fun isRunningOnUiThread(): Boolean =
+    Looper.getMainLooper().thread == Thread.currentThread()